Extracted out block status into it's own component
